Faster Instance Deallocation - overcome VB design
When VB* terminates a class instance, it traverses all instances starting from the last created instance all the way to the first. On itself this would be fast but unfortunately VB also makes checks and can reclaim memory that is unrelated to the instance being terminated. Because of those extra checks, this traversal becomes exponentially slower the more instances there are.
This update 'tricks' VB into traversing only a handful of instances thus making the whole termination process linear and thus much faster.